move maven-resolver stuff to gshell-repository extension
