+++ pom.xml
@@ -392,6 +392,12 @@
 
       <dependency>
         <groupId>com.planet57.gshell.ext</groupId>
+        <artifactId>gshell-repository</artifactId>
+        <version>3.0.0-SNAPSHOT</version>
+      </dependency>
+
+      <dependency>
+        <groupId>com.planet57.gshell.ext</groupId>
         <artifactId>gshell-gossip</artifactId>
         <version>3.0.0-SNAPSHOT</version>
       </dependency>
